pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.11"
discord = "^2.3.2"
prettytable = "^3.8.0"
aiosqlite = "^0.19.0"
jishaku = "^2.5.1"
openai = "^0.27.8"
pillow = "^10.0.0"
psutil = "^5.9.5"
wavelink = "^2.6.3"
aioconsole = "^0.6.2"
pytz = "^2023.3"
aiohttp = "^3.8.5"
humanize = "^4.8.0"
timestamp = "^0.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"